Create order before processing payment
- Orders need to be created before being sent to Stripe for payment.
- Created a migration to add an address foreign key to orders.
- Addresses belonging to orders do not necessarily have to have a
customer attached - these are addresses that are "unsaved", and will not
show up in the customer's addresses during checkout. They do, however,
have an association with the customer through the order.
- Created a migration to remove the not null constraint on the address
table for the customer foreign key.
- Created an orders model and added associations to customers, line
items, payments and addresses.
- Created an orders helper to convert params coming from the front end
into Orders, before being sent to stripe.
- Modified the Payments controller to retrieve a newly-created order
before creating the Stripe session.
- Modified the Stripe helper to create a session from and order.
- Added the tax association to the LineItem model.
- Modified the Photo and Package associations to LineItems to support
the polymorphic relationship.
- Since authentication is still busted, I needed to hack the Orders
helper to supply a current user. THIS NEEDS TO BE FIXED!!